+++ crates/query-engine/metadata/src/metadata/native_queries.rs
@@ -5,6 +5,7 @@ use super::database::*;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
+use std::fs;
 
 // Types
 
@@ -20,7 +21,7 @@ pub struct NativeQueryInfo {
     /// SQL expression to use for the Native Query.
     /// We can interpolate values using `{{variable_name}}` syntax,
     /// such as `SELECT * FROM authors WHERE name = {{author_name}}`
-    pub sql: NativeQuerySql,
+    pub sql: NativeQuerySqlEither,
     /// Columns returned by the Native Query
     pub columns: BTreeMap<String, ReadOnlyColumnInfo>,
     #[serde(default)]
@@ -46,6 +47,149 @@ pub struct ReadOnlyColumnInfo {
     pub description: Option<String>,
 }
 
+/// This type contains information that still needs to be resolved.
+/// After deserializing, we expect the value to be "external",
+/// and after a subsequent step where we read from files,
+/// they should all be converted to NativeQuerySql.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[serde(from = "NativeQuerySqlExternal")]
+#[serde(into = "NativeQuerySqlExternal")]
+pub enum NativeQuerySqlEither {
+    NativeQuerySql(NativeQuerySql),
+    NativeQuerySqlExternal(NativeQuerySqlExternal),
+}
+
+impl NativeQuerySqlEither {
+    /// Extract the actual native query sql from this type.
+    /// If this happens before reading a file, it will fail with an error.
+    pub fn sql(self) -> Result<NativeQueryParts, String> {
+        match self {
+            NativeQuerySqlEither::NativeQuerySql(NativeQuerySql::Inline { sql }) => Ok(sql),
+            NativeQuerySqlEither::NativeQuerySql(NativeQuerySql::FromFile { sql, .. }) => Ok(sql),
+            NativeQuerySqlEither::NativeQuerySqlExternal(NativeQuerySqlExternal::Inline {
+                inline,
+            }) => Ok(inline),
+            NativeQuerySqlEither::NativeQuerySqlExternal(
+                NativeQuerySqlExternal::InlineUntagged(inline),
+            ) => Ok(inline),
+            NativeQuerySqlEither::NativeQuerySqlExternal(NativeQuerySqlExternal::File {
+                ..
+            }) => Err("not all native query sql files were read during parsing".to_string()),
+        }
+    }
+}
+
+impl From<NativeQuerySqlExternal> for NativeQuerySqlEither {
+    /// We use this to deserialize.
+    fn from(value: NativeQuerySqlExternal) -> Self {
+        NativeQuerySqlEither::NativeQuerySqlExternal(value)
+    }
+}
+
+impl From<NativeQuerySqlEither> for NativeQuerySqlExternal {
+    /// We use this to serialize.
+    fn from(value: NativeQuerySqlEither) -> Self {
+        match value {
+            NativeQuerySqlEither::NativeQuerySqlExternal(value) => value,
+            NativeQuerySqlEither::NativeQuerySql(value) => value.into(),
+        }
+    }
+}
+
+/// A Native Query SQL after file resolution.
+/// This is the underlying type of the `NativeQuerySqlEither` variant with the same name
+/// that is expected in the metadata when translating requests. A subsequent phase after de-serializing
+/// Should convert NativeQuerySqlExternal values to values of this type.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum NativeQuerySql {
+    FromFile {
+        file: std::path::PathBuf,
+        sql: NativeQueryParts,
+    },
+    Inline {
+        sql: NativeQueryParts,
+    },
+}
+
+impl NativeQuerySql {
+    /// Extract the native query sql expression.
+    pub fn sql(self) -> NativeQueryParts {
+        match self {
+            NativeQuerySql::Inline { sql } => sql,
+            NativeQuerySql::FromFile { sql, .. } => sql,
+        }
+    }
+}
+
+// We use this type as an intermediate representation for serialization/deserialization
+// of native query sql location/expression.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "camelCase")]
+#[serde(untagged)]
+/// Native Query SQL location.
+pub enum NativeQuerySqlExternal {
+    /// Refer to an external Native Query SQL file.
+    File {
+        /// Relative path to a sql file.
+        file: std::path::PathBuf,
+    },
+    /// Inline Native Query SQL string.
+    Inline {
+        /// An inline Native Query SQL string.
+        inline: NativeQueryParts,
+    },
+    InlineUntagged(
+        /// An inline Native Query SQL string.
+        NativeQueryParts,
+    ),
+}
+
+impl NativeQuerySqlEither {
+    /// Convert an external native query sql type to NativeQuerySql,
+    /// including reading files from disk.
+    pub fn from_external(
+        &self,
+        absolute_configuration_directory: &std::path::Path,
+    ) -> Result<NativeQuerySql, String> {
+        match self {
+            // unexpected we get this, but ok.
+            NativeQuerySqlEither::NativeQuerySql(value) => Ok(value.clone()),
+            NativeQuerySqlEither::NativeQuerySqlExternal(external) => match external {
+                NativeQuerySqlExternal::File { file } => {
+                    parse_native_query_from_file(absolute_configuration_directory.join(file))
+                }
+                NativeQuerySqlExternal::Inline { inline } => Ok(NativeQuerySql::Inline {
+                    sql: inline.clone(),
+                }),
+                NativeQuerySqlExternal::InlineUntagged(inline) => Ok(NativeQuerySql::Inline {
+                    sql: inline.clone(),
+                }),
+            },
+        }
+    }
+}
+
+impl From<NativeQuerySql> for NativeQuerySqlExternal {
+    /// used for deserialization.
+    fn from(value: NativeQuerySql) -> Self {
+        match value {
+            NativeQuerySql::Inline { sql } => NativeQuerySqlExternal::Inline { inline: sql },
+            NativeQuerySql::FromFile { file, .. } => NativeQuerySqlExternal::File { file },
+        }
+    }
+}
+
+impl JsonSchema for NativeQuerySqlEither {
+    fn schema_name() -> String {
+        "NativeQuerySql".to_string()
+    }
+
+    fn json_schema(gen: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
+        NativeQuerySqlExternal::json_schema(gen)
+    }
+}
+
 /// A part of a Native Query text, either raw text or a parameter.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum NativeQueryPart {
@@ -55,19 +199,24 @@ pub enum NativeQueryPart {
     Parameter(String),
 }
 
-/// A Native Query SQL after parsing.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct NativeQuerySql(pub Vec<NativeQueryPart>);
+/// A Native Query SQL parts after parsing.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(from = "String")]
+#[serde(into = "String")]
+pub struct NativeQueryParts(pub Vec<NativeQueryPart>);
 
-// Serialization
+impl From<String> for NativeQueryParts {
+    /// Used for de-serialization.
+    fn from(value: String) -> Self {
+        parse_native_query(&value)
+    }
+}
 
-impl Serialize for NativeQuerySql {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
+impl From<NativeQueryParts> for String {
+    /// Used for serialization.
+    fn from(value: NativeQueryParts) -> Self {
         let mut sql: String = String::new();
-        for part in self.0.iter() {
+        for part in value.0.iter() {
             match part {
                 NativeQueryPart::Text(text) => sql.push_str(text.as_str()),
                 NativeQueryPart::Parameter(param) => {
@@ -75,42 +224,13 @@ impl Serialize for NativeQuerySql {
                 }
             }
         }
-        serializer.serialize_str(&sql)
+        sql
     }
 }
 
-struct NQVisitor;
-
-impl<'de> serde::de::Visitor<'de> for NQVisitor {
-    type Value = NativeQuerySql;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-        formatter.write_str("A Native Query SQL")
-    }
-
-    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-    where
-        E: serde::de::Error,
-    {
-        // todo: write a better parser
-        let sql = parse_native_query(value);
-
-        Ok(NativeQuerySql(sql))
-    }
-}
-
-impl<'de> Deserialize<'de> for NativeQuerySql {
-    fn deserialize<D>(deserializer: D) -> Result<NativeQuerySql, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        deserializer.deserialize_str(NQVisitor)
-    }
-}
-
-impl JsonSchema for NativeQuerySql {
+impl JsonSchema for NativeQueryParts {
     fn schema_name() -> String {
-        "Native_query_sql".to_string()
+        "InlineNativeQuerySql".to_string()
     }
 
     fn json_schema(gen: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
@@ -120,9 +240,18 @@ impl JsonSchema for NativeQuerySql {
 
 // Parsing
 
-/// Parse a native query into parts where variables have the
-/// syntax `{{<variable>}}`.
-fn parse_native_query(string: &str) -> Vec<NativeQueryPart> {
+/// Read a file a parse it into native query parts.
+pub fn parse_native_query_from_file(file: std::path::PathBuf) -> Result<NativeQuerySql, String> {
+    let contents: String = match fs::read_to_string(&file) {
+        Ok(ok) => Ok(ok),
+        Err(err) => Err(format!("{}: {}", &file.display(), err)),
+    }?;
+    let sql = parse_native_query(&contents);
+    Ok(NativeQuerySql::FromFile { file, sql })
+}
+
+/// Parse a native query into parts where variables have the syntax `{{<variable>}}`.
+fn parse_native_query(string: &str) -> NativeQueryParts {
     let vec: Vec<Vec<NativeQueryPart>> = string
         .split("{{")
         .map(|part| match part.split_once("}}") {
@@ -139,18 +268,20 @@ fn parse_native_query(string: &str) -> Vec<NativeQueryPart> {
             }
         })
         .collect();
-    vec.concat()
+    NativeQueryParts(vec.concat())
 }
+
+// tests
 
 #[cfg(test)]
 mod tests {
-    use super::{parse_native_query, NativeQueryPart};
+    use super::{parse_native_query, NativeQueryPart, NativeQueryParts, NativeQuerySqlExternal};
 
     #[test]
     fn no_parameters() {
         assert_eq!(
             parse_native_query("select 1"),
-            vec![NativeQueryPart::Text("select 1".to_string())]
+            NativeQueryParts(vec![NativeQueryPart::Text("select 1".to_string())])
         );
     }
 
@@ -158,11 +289,11 @@ mod tests {
     fn one_parameter() {
         assert_eq!(
             parse_native_query("select * from t where {{name}} = name"),
-            vec![
+            NativeQueryParts(vec![
                 NativeQueryPart::Text("select * from t where ".to_string()),
                 NativeQueryPart::Parameter("name".to_string()),
                 NativeQueryPart::Text(" = name".to_string()),
-            ]
+            ])
         );
     }
 
@@ -170,14 +301,14 @@ mod tests {
     fn multiple_parameters() {
         assert_eq!(
             parse_native_query("select * from t where id = {{id}} and {{name}} = {{other_name}}"),
-            vec![
+            NativeQueryParts(vec![
                 NativeQueryPart::Text("select * from t where id = ".to_string()),
                 NativeQueryPart::Parameter("id".to_string()),
                 NativeQueryPart::Text(" and ".to_string()),
                 NativeQueryPart::Parameter("name".to_string()),
                 NativeQueryPart::Text(" = ".to_string()),
                 NativeQueryPart::Parameter("other_name".to_string()),
-            ]
+            ])
         );
     }
 
@@ -185,11 +316,31 @@ mod tests {
     fn one_parameter_and_curly_text() {
         assert_eq!(
             parse_native_query("select * from t where {{name}} = '{name}'"),
-            vec![
+            NativeQueryParts(vec![
                 NativeQueryPart::Text("select * from t where ".to_string()),
                 NativeQueryPart::Parameter("name".to_string()),
                 NativeQueryPart::Text(" = '{name}'".to_string()),
-            ]
+            ])
+        );
+    }
+
+    #[test]
+    fn parse_inline_untagged() {
+        assert_eq!(
+            serde_json::from_str::<NativeQuerySqlExternal>(r#""select 1""#).unwrap(),
+            NativeQuerySqlExternal::InlineUntagged(NativeQueryParts(vec![NativeQueryPart::Text(
+                "select 1".to_string()
+            )]))
+        );
+    }
+
+    #[test]
+    fn parse_inline_tagged() {
+        assert_eq!(
+            serde_json::from_str::<NativeQuerySqlExternal>(r#"{ "inline": "select 1" }"#).unwrap(),
+            NativeQuerySqlExternal::Inline {
+                inline: NativeQueryParts(vec![NativeQueryPart::Text("select 1".to_string())])
+            }
         );
     }
 }